refactor: Extract CodeHighlight to own file
