+++ src/components/CodeBlock.js
@@ -1,49 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Button from './Button';
-import Highlight from 'prism-react-renderer';
+import CodeHighlight from './CodeHighlight';
 import FeatherIcon from './FeatherIcon';
 import MiddleEllipsis from 'react-middle-ellipsis';
-import Prism from 'prismjs';
 import styles from './CodeBlock.module.scss';
 import useClipboard from '../hooks/useClipboard';
 import useFormattedCode from '../hooks/useFormattedCode';
-
-const CodeHighlight = ({ children, language, lineNumbers }) => {
-  return (
-    <Highlight Prism={Prism} code={children.trim()} language={language}>
-      {({ tokens, getLineProps, getTokenProps }) => {
-        const characterWidth = String(tokens.length).length;
-
-        return (
-          <pre className={styles.codeContainer} data-language={language}>
-            <code>
-              {tokens.map((line, i) => (
-                // eslint-disable-next-line react/jsx-key
-                <div {...getLineProps({ line, key: i })}>
-                  {lineNumbers && (
-                    <span
-                      className={styles.lineNumber}
-                      style={{
-                        '--character-width': `${characterWidth}ch`,
-                      }}
-                    >
-                      {i + 1}
-                    </span>
-                  )}
-                  {line.map((token, key) => (
-                    // eslint-disable-next-line react/jsx-key
-                    <span {...getTokenProps({ token, key })} />
-                  ))}
-                </div>
-              ))}
-            </code>
-          </pre>
-        );
-      }}
-    </Highlight>
-  );
-};
 
 const CodeBlock = ({ children, copy, fileName, language, lineNumbers }) => {
   const formattedCode = useFormattedCode(children.trim());
@@ -78,12 +41,6 @@ const CodeBlock = ({ children, copy, fileName, language, lineNumbers }) => {
       )}
     </div>
   );
-};
-
-CodeHighlight.propTypes = {
-  children: PropTypes.string.isRequired,
-  language: PropTypes.string,
-  lineNumbers: PropTypes.bool,
 };
 
 CodeBlock.propTypes = {
